transform load-test in tests
